Silence warnings and shorten sleep duration

+++ CircuitPlayground/toolbox.cpp
@@ -119,7 +119,7 @@ void Toolbox::processMouseMotionEvent(const SDL_MouseMotionEvent& event) {
     if (index >= MainWindow::tool_tags::size) return;
 
     // check if the mouse is on the button spacing instead of on the actual button
-    if ((offsetY - PADDING_VERTICAL) - index * (BUTTON_HEIGHT + BUTTON_SPACING) >= BUTTON_HEIGHT) return;
+    if ((offsetY - PADDING_VERTICAL) - static_cast<int32_t>(index) * (BUTTON_HEIGHT + BUTTON_SPACING) >= BUTTON_HEIGHT) return;
 
     // save the index since it is valid
     mouseoverToolIndex = index;
@@ -139,7 +139,7 @@ void Toolbox::processMouseButtonDownEvent(const SDL_MouseButtonEvent& event) {
     if (index >= MainWindow::tool_tags::size) return;
 
     // check if the mouse is on the button spacing instead of on the actual button
-    if ((offsetY - PADDING_VERTICAL) - index * (BUTTON_HEIGHT + BUTTON_SPACING) >= BUTTON_HEIGHT) return;
+    if ((offsetY - PADDING_VERTICAL) - static_cast<int32_t>(index) * (BUTTON_HEIGHT + BUTTON_SPACING) >= BUTTON_HEIGHT) return;
 
     // if there is already a handle binded to this tool, then swap the handles
     auto it = std::find(mainWindow.selectedToolIndices, mainWindow.selectedToolIndices + NUM_INPUT_HANDLES, index);
